+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - Use remote memory access to collect finished visibility data, without serialization.
 
 ### Fixed
+- Corrects the distribution of random points for the random mock catalog.
 - pixel interpolation was defaulting to az_za_simple for all beams, breaking healpix-coord UVBeams.
 
 ## [1.2.0] - 2020-7-20